Feat setup database functions
Make Reservation: insert a reservation, post information to the handler, pass it off to a database function and writ that reservation to the database.

the user might lose his or her connection to the Internet partway through the transaction. take advantage of the context to make sure that this transaction doesn't stay open for the three-second lifetime that we have for transactions if nothing is happening.

return an error flag, and the id of last inserted record

build up RoomRestriction model using the id of last reservation posted