Feat Send Emails
send two email messages
send one to the property owner just to alert him or her that there is a new reservation. And I want to send a confirmation email to the person who made the reservation in the first place.

in order to send email to need to know certain things at a bare minimum.
the to address, the from address, the subject, and the content.
use channels to send emails
put MailChan in AppConfig to be available in the whole application
create a chan, defer close the channel when the application stops
implement the sending a message

listenForMail() starts to listen for Mail
runs indefinitely and fires things off in the background to happen asynchronously.
then the constructed msg data sent to the MailChan(channel)
this channel pass the msg to sendMsg()

sendMsg()
connect to the mail server and client
to write this procedurally and just send constructive message, connect to the mail server, send the message and keep going